FINAL VERSION OF 5.3-5.4

+++ iv1201Group12/src/main/java/kth/iv1201/Group12/domain/CompetenceProfileDTO.java
@@ -4,7 +4,7 @@ public class CompetenceProfileDTO {
     private int competenceId;
     private float yearsOfExperience;
 
-    public CompetenceProfileDTO(int competenceId, int yearsOfExperience) {
+    public CompetenceProfileDTO(int competenceId,float yearsOfExperience) {
         this.competenceId = competenceId;
         this.yearsOfExperience = yearsOfExperience;
     }
